pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dynamic-crypto-pricing"
version = "0.1.0"
description = "Dynamic pricing engine for merchants accepting cryptocurrencies"
authors = [{ name = "Your Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "requests>=2.31.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]
